# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Simple API Restaurant",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/meals": {
            "get": {
                "tags": [
                    "meals"
                ],
                "summary": "List all meals",
                "description": "Return a list of meals",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Please provide correct data"
                    }
                }
            },
            "post": {
                "tags": [
                    "meals"
                ],
                "summary": "Create one meal",
                "description": "Create a meal resource",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Title of meal",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description of meal",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity, validation fails."
                    }
                }
            }
        },
        "/meals/{meal}": {
            "get": {
                "tags": [
                    "meals"
                ],
                "summary": "Read one meal",
                "description": "Return a meal resource",
                "parameters": [
                    {
                        "name": "meal",
                        "in": "path",
                        "description": "Id of meal",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Please provide correct data"
                    }
                }
            },
            "put": {
                "tags": [
                    "meals"
                ],
                "summary": "Update one meal",
                "description": "Update a meal and return a meal name and description.",
                "parameters": [
                    {
                        "name": "meal",
                        "in": "path",
                        "description": "Id of meal",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Mame of meal",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description of meal",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity, validation fails."
                    }
                }
            },
            "delete": {
                "tags": [
                    "meals"
                ],
                "summary": "Delete one meal.",
                "description": "A meal resource will be deleted",
                "parameters": [
                    {
                        "name": "meal",
                        "in": "path",
                        "description": "Id of meal",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resource deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
}